Initial commit.
I am terribly sorry (most especially for myself) to have done this so
late into the project that the initial commit has basically the final
implementation of all files. Hoenstly, I feel bad for this because I
love tracking my history of changes. It's been 3 days so far, and maybe
3 days isn't much to look back at but it's something I wish I kept.

Alas, there is no undoing.

Anyhow, here is the "initial" (haha) commit for th entire project.