Added more implementation to the Tester file.
I also included merge sort as part of the tester class because we need
it to compute the entropy of a shuffle. Note that we do not use merge
sort to verify sortedness (we do this without any sorting implementations),
but we use it to compute stuff about shuffled records.

The primary thing added were helper function for the most part (_Tester_copy(),
Tester_exit(), and Tester_recordsFill()).