Added implementation for computing r squared; verified sort checker.
The coefficient of correlation can tell us whether or not two pieces of data
have some form of correlation. Now if the indices of the shuffled records are
correlated with the indices of the sorted records, then most likely some sort
of pattern exists in the shuffle (it isn't well shuffled). Thus, lower R values
are good for telling us that a deck is well shuffled (no apparent order). Squaring
R means the order doesn't matter (a list in reverse is considered not well shuffled,
which makes sense, even if it isn't "sorted").