Edited README and fixed random.c subtlety.
My initial intuition was wrong. Simply putting <= would not work because if p=0
and the chosen number happened to be 0, the function would return 1 (which
violates the assumption that p=0 means the function would always return 0).
I fixed the logic to account for this. Now p=0 always returns 0 and p=1
always returns 1.