You won't believe what bug I came across...
So my smooth sort implementation was working completely fine on Linux,
when suddenly, after running it on Windows, it completely failed 100%
of the time. I was so worried and wasn't sure what might be wrong with
my implementation; I double checked and ran the most recent code on Linux,
but it was working fine there. That's when I was like "holy sht! it's a bug
caused by platform differences??" Now that doesn't happen often.

The best part about this bug was that it had nothing to do with my code...
when performing smooth sort, I used log(num) / log(2) to compute the
logarithm of a number in base 2... this worked completely fine on Linux,
but on Windows, for some fucking reason, the precision broke down and
log(8) / log(2) = 2! (This also produced the incorrect result for other
powers of 2)... So I had to migrate and use log2() instead, which I had no
idea existed.